# api/openapi/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Keystone Core REST API
  description: |
    REST API for Keystone Core - a runtime infrastructure control plane.

    This specification covers all HTTP REST endpoints exposed by Keystone Core components.
    The primary control plane operations use gRPC (see proto files), but HTTP REST endpoints
    are available for cluster management, health checks, file operations, and more.

    ## Authentication

    Keystone Core supports multiple authentication methods:
    - **API Key**: Bearer token in `Authorization` header
    - **mTLS**: Client certificate authentication
    - **JWT**: JSON Web Token authentication

    ## Rate Limiting

    API endpoints are subject to rate limiting. When rate limited, the API returns
    HTTP 429 with `Retry-After` header indicating when to retry.

    ## Versioning

    All REST APIs use the `/api/v1/` prefix. Breaking changes will increment the version.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: Keystone Core Team
    url: https://github.com/shawnbutts/keystone-core
servers:
  - url: http://localhost:9090
    description: Local development server
  - url: https://{control-plane-host}:9443
    description: Production server (HTTPS)
    variables:
      control-plane-host:
        default: control-plane.example.com
        description: Control plane hostname
...
security:
  - bearerAuth: []
  - mtls: []
paths:
...
  /api/v1/state/apply:
    post:
      tags:
        - State
      summary: Apply state
      description: |
        Applies a declarative state definition. Supports dry-run mode to preview
        changes without applying them.
      operationId: applyState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateRequest'
      responses:
        '200':
          description: Apply result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/state/check:
    post:
      tags:
        - State
      summary: Validate state
      description: Validates a state definition without applying it.
      operationId: checkState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/state/drift:
    post:
      tags:
        - State
      summary: Check drift
      description: Checks for configuration drift against a state definition.
      operationId: checkDrift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateRequest'
      responses:
        '200':
          description: Drift report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # Maintenance Mode
  # ============================================
  /api/v1/maintenance/enable:
    post:
      tags:
        - Maintenance
      summary: Enable maintenance mode
      description: Enables maintenance mode, queuing incoming commands.
      operationId: enableMaintenance
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceEnableRequest'
      responses:
        '200':
          description: Maintenance enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
                  reason:
                    type: string
                  since:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Already in maintenance mode (use force=true to override)
  /api/v1/maintenance/disable:
    post:
      tags:
        - Maintenance
      summary: Disable maintenance mode
      description: Disables maintenance mode and releases queued commands.
      operationId: disableMaintenance
      responses:
        '200':
          description: Maintenance disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
                  disabled_at:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Maintenance mode is not enabled
  /api/v1/maintenance/status:
    get:
      tags:
        - Maintenance
      summary: Get maintenance status
      description: Returns current maintenance mode status.
      operationId: getMaintenanceStatus
      responses:
        '200':
          description: Maintenance status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/maintenance/queue:
    get:
      tags:
        - Maintenance
      summary: Get maintenance queue
      description: Returns commands queued during maintenance mode.
      operationId: getMaintenanceQueue
      responses:
        '200':
          description: Queued commands
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceQueue'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/maintenance/cleanup:
    post:
      tags:
        - Maintenance
      summary: Cleanup maintenance data
      description: Cleans up old maintenance records.
      operationId: cleanupMaintenance
      parameters:
        - name: older_than
          in: query
          description: Delete records older than this duration
          schema:
            type: string
        - name: dry_run
          in: query
          description: Preview cleanup without deleting
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Cleanup result
          content:
            application/json:
              schema:
                type: object
                properties:
                  dry_run:
                    type: boolean
                  older_than:
                    type: string
                  deleted:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # API Keys
  # ============================================
...
  /api/v1/webhooks/stats:
    get:
      tags:
        - Webhooks
      summary: Webhook statistics
      description: |
        Returns webhook receiver statistics.
        Note: Handler exists but is not yet wired into kscore-server.
      operationId: getWebhookStats
      responses:
        '200':
          description: Webhook statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key authentication
    mtls:
      type: mutualTLS
      description: Mutual TLS client certificate authentication
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
...
    StateRequest:
      type: object
      properties:
        content:
          type: string
          description: State file content (YAML)
        path:
          type: string
          description: State file path (local on server)
        vars:
          type: object
          description: Variables to pass to templates
        dry_run:
          type: boolean
          description: Preview changes without applying
        targets:
          type: array
          items:
            type: string
          description: Target specific states by ID
    ApplyResponse:
      type: object
      properties:
        run_id:
          type: string
        status:
          type: string
          enum: [success, changed, failed, unknown]
        summary:
          $ref: '#/components/schemas/StateSummary'
        results:
          type: array
          items:
            $ref: '#/components/schemas/StateResult'
        started_at:
          type: string
          format: date-time
        duration:
          type: string
        dry_run:
          type: boolean
    StateSummary:
      type: object
      properties:
        total:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        changed:
          type: integer
        unchanged:
          type: integer
        skipped:
          type: integer
    StateResult:
      type: object
      properties:
        id:
          type: string
        module:
          type: string
        status:
          type: string
          enum: [success, failed]
        changed:
          type: boolean
        comment:
          type: string
        error:
          type: string
        duration:
          type: string
    CheckResponse:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        states:
          type: integer
        modules:
          type: array
          items:
            type: string
    ValidationError:
      type: object
      properties:
        state_id:
          type: string
        field:
          type: string
        message:
          type: string
        line:
          type: integer
    DriftResponse:
      type: object
      properties:
        run_id:
          type: string
        has_drift:
          type: boolean
        summary:
          $ref: '#/components/schemas/DriftSummary'
        states:
          type: array
          items:
            $ref: '#/components/schemas/DriftState'
        checked_at:
          type: string
          format: date-time
        duration:
          type: string
    DriftSummary:
      type: object
      properties:
        total:
          type: integer
        no_drift:
          type: integer
        low:
          type: integer
        medium:
          type: integer
        high:
          type: integer
        critical:
          type: integer
    DriftState:
      type: object
      properties:
        state_id:
          type: string
        module:
          type: string
        has_drift:
          type: boolean
        severity:
          type: string
          enum: [none, low, medium, high, critical]
        differences:
          type: array
          items:
            $ref: '#/components/schemas/DriftDiff'
    DriftDiff:
      type: object
      properties:
        path:
          type: string
        expected:
          description: Expected value
        actual:
          description: Actual value
        severity:
          type: string
        message:
          type: string
    MaintenanceEnableRequest:
      type: object
      properties:
        reason:
          type: string
        duration:
          type: string
          description: Duration string (e.g., "1h", "30m")
        force:
          type: boolean
          default: false
    MaintenanceStatus:
      type: object
      properties:
        enabled:
          type: boolean
        reason:
          type: string
        since:
          type: string
          format: date-time
        duration:
          type: string
        enabled_by:
          type: string
    MaintenanceQueue:
      type: object
      properties:
        queued_commands:
          type: integer
        commands:
          type: array
          items:
            type: object
...
    WebhookStats:
      type: object
      properties:
        totalReceived:
          type: integer
          format: int64
        totalProcessed:
          type: integer
          format: int64
        totalFailed:
          type: integer
          format: int64
        bySource:
          type: object
          additionalProperties:
            type: object
            properties:
              received:
                type: integer
                format: int64
              processed:
                type: integer
                format: int64
              failed:
                type: integer
                format: int64
        lastReceived:
          type: string
          format: date-time
...